fixes direction fuckup

+++ app/controllers/drivers_controller.rb
@@ -3,6 +3,6 @@ class DriversController < ApplicationController
   before_filter :authenticate_user!
 
   def index
-    @drivers = Driver.within_bounding_box(get_bounding_box(current_user, params[:search_distance].to_i))
+    @drivers = Driver.within_bounding_box(get_bounding_box(current_user, params[:search_distance].to_f))
   end
 end
